chore: split components apart
